+++ recipe/session/sessionwithjwt/sessionContainer.go
@@ -27,44 +27,44 @@ import (
 func newSessionWithJWTContainer(originalSessionClass sessmodels.SessionContainer, openidRecipeImplementation openidmodels.RecipeInterface) sessmodels.SessionContainer {
 
 	return sessmodels.SessionContainer{
-		RevokeSession: func(userContext supertokens.UserContext) error {
-			return originalSessionClass.RevokeSession(userContext)
+		RevokeSessionWithContext: func(userContext supertokens.UserContext) error {
+			return originalSessionClass.RevokeSessionWithContext(userContext)
 		},
 
-		GetSessionData: func(userContext supertokens.UserContext) (map[string]interface{}, error) {
-			return originalSessionClass.GetSessionData(userContext)
+		GetSessionDataWithContext: func(userContext supertokens.UserContext) (map[string]interface{}, error) {
+			return originalSessionClass.GetSessionDataWithContext(userContext)
 		},
 
-		UpdateSessionData: func(newSessionData map[string]interface{}, userContext supertokens.UserContext) error {
-			return originalSessionClass.UpdateSessionData(newSessionData, userContext)
+		UpdateSessionDataWithContext: func(newSessionData map[string]interface{}, userContext supertokens.UserContext) error {
+			return originalSessionClass.UpdateSessionDataWithContext(newSessionData, userContext)
 		},
-		GetUserID: func(userContext supertokens.UserContext) string {
-			return originalSessionClass.GetUserID(userContext)
+		GetUserIDWithContext: func(userContext supertokens.UserContext) string {
+			return originalSessionClass.GetUserIDWithContext(userContext)
 		},
-		GetAccessTokenPayload: func(userContext supertokens.UserContext) map[string]interface{} {
-			return originalSessionClass.GetAccessTokenPayload(userContext)
+		GetAccessTokenPayloadWithContext: func(userContext supertokens.UserContext) map[string]interface{} {
+			return originalSessionClass.GetAccessTokenPayloadWithContext(userContext)
 		},
-		GetHandle: func(userContext supertokens.UserContext) string {
-			return originalSessionClass.GetHandle(userContext)
+		GetHandleWithContext: func(userContext supertokens.UserContext) string {
+			return originalSessionClass.GetHandleWithContext(userContext)
 		},
-		GetAccessToken: func(userContext supertokens.UserContext) string {
-			return originalSessionClass.GetAccessToken(userContext)
+		GetAccessTokenWithContext: func(userContext supertokens.UserContext) string {
+			return originalSessionClass.GetAccessTokenWithContext(userContext)
 		},
-		GetTimeCreated: func(userContext supertokens.UserContext) (uint64, error) {
-			return originalSessionClass.GetTimeCreated(userContext)
+		GetTimeCreatedWithContext: func(userContext supertokens.UserContext) (uint64, error) {
+			return originalSessionClass.GetTimeCreatedWithContext(userContext)
 		},
-		GetExpiry: func(userContext supertokens.UserContext) (uint64, error) {
-			return originalSessionClass.GetExpiry(userContext)
+		GetExpiryWithContext: func(userContext supertokens.UserContext) (uint64, error) {
+			return originalSessionClass.GetExpiryWithContext(userContext)
 		},
-		UpdateAccessTokenPayload: func(newAccessTokenPayload map[string]interface{}, userContext supertokens.UserContext) error {
+		UpdateAccessTokenPayloadWithContext: func(newAccessTokenPayload map[string]interface{}, userContext supertokens.UserContext) error {
 			if newAccessTokenPayload == nil {
 				newAccessTokenPayload = map[string]interface{}{}
 			}
-			accessTokenPayload := originalSessionClass.GetAccessTokenPayload(userContext)
+			accessTokenPayload := originalSessionClass.GetAccessTokenPayloadWithContext(userContext)
 			jwtPropertyName, ok := accessTokenPayload[ACCESS_TOKEN_PAYLOAD_JWT_PROPERTY_NAME_KEY]
 
 			if !ok {
-				return originalSessionClass.UpdateAccessTokenPayload(newAccessTokenPayload, userContext)
+				return originalSessionClass.UpdateAccessTokenPayloadWithContext(newAccessTokenPayload, userContext)
 			}
 
 			existingJWT := accessTokenPayload[jwtPropertyName.(string)].(string)
@@ -91,12 +91,83 @@ func newSessionWithJWTContainer(originalSessionClass sessmodels.SessionContainer
 				jwtExpiry = 1
 			}
 
-			newAccessTokenPayload, err = addJWTToAccessTokenPayload(newAccessTokenPayload, jwtExpiry, originalSessionClass.GetUserID(userContext), jwtPropertyName.(string), openidRecipeImplementation, &map[string]interface{}{})
+			newAccessTokenPayload, err = addJWTToAccessTokenPayload(newAccessTokenPayload, jwtExpiry, originalSessionClass.GetUserIDWithContext(userContext), jwtPropertyName.(string), openidRecipeImplementation, &map[string]interface{}{})
 			if err != nil {
 				return err
 			}
 
-			return originalSessionClass.UpdateAccessTokenPayload(newAccessTokenPayload, userContext)
+			return originalSessionClass.UpdateAccessTokenPayloadWithContext(newAccessTokenPayload, userContext)
+		},
+		RevokeSession: func() error {
+			return originalSessionClass.RevokeSessionWithContext(nil)
+		},
+
+		GetSessionData: func() (map[string]interface{}, error) {
+			return originalSessionClass.GetSessionDataWithContext(nil)
+		},
+
+		UpdateSessionData: func(newSessionData map[string]interface{}) error {
+			return originalSessionClass.UpdateSessionDataWithContext(newSessionData, nil)
+		},
+		GetUserID: func() string {
+			return originalSessionClass.GetUserIDWithContext(nil)
+		},
+		GetAccessTokenPayload: func() map[string]interface{} {
+			return originalSessionClass.GetAccessTokenPayloadWithContext(nil)
+		},
+		GetHandle: func() string {
+			return originalSessionClass.GetHandleWithContext(nil)
+		},
+		GetAccessToken: func() string {
+			return originalSessionClass.GetAccessTokenWithContext(nil)
+		},
+		GetTimeCreated: func() (uint64, error) {
+			return originalSessionClass.GetTimeCreatedWithContext(nil)
+		},
+		GetExpiry: func() (uint64, error) {
+			return originalSessionClass.GetExpiryWithContext(nil)
+		},
+		UpdateAccessTokenPayload: func(newAccessTokenPayload map[string]interface{}) error {
+			if newAccessTokenPayload == nil {
+				newAccessTokenPayload = map[string]interface{}{}
+			}
+			accessTokenPayload := originalSessionClass.GetAccessTokenPayloadWithContext(nil)
+			jwtPropertyName, ok := accessTokenPayload[ACCESS_TOKEN_PAYLOAD_JWT_PROPERTY_NAME_KEY]
+
+			if !ok {
+				return originalSessionClass.UpdateAccessTokenPayloadWithContext(newAccessTokenPayload, nil)
+			}
+
+			existingJWT := accessTokenPayload[jwtPropertyName.(string)].(string)
+
+			currentTimeInSeconds := uint64(time.Now().UnixNano() / 1000000000) // time in seconds
+
+			claims := jwt.MapClaims{}
+			decodedPayload := map[string]interface{}{}
+			_, _, err := new(jwt.Parser).ParseUnverified(existingJWT, claims)
+			if err != nil {
+				return err
+			}
+			for key, val := range claims {
+				decodedPayload[key] = val
+			}
+
+			jwtExpiry := uint64(decodedPayload["exp"].(float64)) - currentTimeInSeconds
+
+			if jwtExpiry <= 0 {
+				// it can come here if someone calls this function well after
+				// the access token and the jwt payload have expired (which can happen if an API takes a VERY long time). In this case, we still want the jwt payload to update, but the resulting JWT should
+				// not be alive for too long (since it's expired already). So we set it to
+				// 1 second lifetime.
+				jwtExpiry = 1
+			}
+
+			newAccessTokenPayload, err = addJWTToAccessTokenPayload(newAccessTokenPayload, jwtExpiry, originalSessionClass.GetUserIDWithContext(nil), jwtPropertyName.(string), openidRecipeImplementation, &map[string]interface{}{})
+			if err != nil {
+				return err
+			}
+
+			return originalSessionClass.UpdateAccessTokenPayloadWithContext(newAccessTokenPayload, nil)
 		},
 	}
 }