+++ test/auth-react-server/main.go
@@ -435,7 +435,7 @@ func sessioninfo(w http.ResponseWriter, r *http.Request) {
 		w.Write([]byte("no session found"))
 		return
 	}
-	sessionData, err := sessionContainer.GetSessionData(nil)
+	sessionData, err := sessionContainer.GetSessionData()
 	if err != nil {
 		err = supertokens.ErrorHandler(err, r, w)
 		if err != nil {
@@ -447,9 +447,9 @@ func sessioninfo(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(200)
 	w.Header().Add("content-type", "application/json")
 	bytes, err := json.Marshal(map[string]interface{}{
-		"sessionHandle":      sessionContainer.GetHandle(nil),
-		"userId":             sessionContainer.GetUserID(nil),
-		"accessTokenPayload": sessionContainer.GetAccessTokenPayload(nil),
+		"sessionHandle":      sessionContainer.GetHandle(),
+		"userId":             sessionContainer.GetUserID(),
+		"accessTokenPayload": sessionContainer.GetAccessTokenPayload(),
 		"sessionData":        sessionData,
 	})
 	if err != nil {